update tests & improve foldouts
