+++ com.unity.render-pipelines.core/Editor/Lighting/ProbeVolume/ProbeReferenceVolumeProfileEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ebad938968750cd41a0d243f50a3a4ba
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
